requirements.txt
Flask==2.3.2
Pillow==9.4.0
docx2pdf==0.4.3
PyMuPDF==1.21.2
pdf2docx==0.8.0
